Moved commands away from main()
