+++ slackbot.go
@@ -17,6 +17,7 @@ limitations under the License.
 package main
 
 import (
+	"context"
 	"fmt"
 	"os"
 	"time"
@@ -26,6 +27,58 @@ import (
 	pagerduty "github.com/PagerDuty/go-pagerduty"
 	"github.com/shomali11/slacker"
 )
+
+// Start will run the main code for the Slack->PD Bot
+func Start(config *config) error {
+	bot := slacker.NewClient(config.slack.token)
+	// defining which function handles the bot Init phase
+	bot.Init(
+		func() {
+			Init(config)
+		})
+
+	// error raised by the Bot will be handled by this function
+	bot.Err(
+		func(err string) {
+			Err(err)
+		})
+
+	// function tied to sentences sent to the Bot and starting with "open emergency" followed by some text
+	bot.Command("open emergency <msg>",
+		"Open an EMERGENCY incident to Customer Support",
+		func(request slacker.Request, response slacker.ResponseWriter) {
+			Emergency(request, response, config)
+		})
+
+	// when no other "Command" matches and text is sent to the Bot, this function will be run instead
+	bot.DefaultCommand(
+		func(request slacker.Request, response slacker.ResponseWriter) {
+			Default(request, response, config)
+		})
+
+	// function run for all events received by the bot (including time ticks)
+	bot.DefaultEvent(
+		func(event interface{}) {
+			//log.Println(event)
+		})
+
+	// set the "help" message handling function
+	bot.Help("help",
+		slacker.WithHandler(
+			func(request slacker.Request, response slacker.ResponseWriter) {
+				help(response, config)
+			}))
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	err := bot.Listen(ctx)
+	if err != nil {
+		Err(err.Error())
+	}
+
+	return nil
+}
 
 // Init is called upon Bot creation (first startup)
 func Init(config *config) {
